Node: slight adjustment to propagation check
Before propagating a message we remove source node of the message from
the list of subscribers because it is not optimal to send message back
to the source from which we just received the message! Adjusted the
condition in which we check if the message was propgated succesfully to
enough nodes to acommodate this fact.